feat(ui): address review feedback and harden dev workflow
- extract theme utils + echarts theme builder from App.tsx

- debounce UI override persistence and improve reset semantics

- avoid unnecessary echarts re-register, remove focus !important, add CSS fallbacks

- add backend-unreachable friendly UX in frontend and startup hint in start_frontend.sh

- add --without-ui-theme to dashboard builder and unit tests

- make hooks/check scripts robust without rg and pycache permission issues
