+++ README.md
@@ -59,10 +59,15 @@ Default ports:
 Start:
 
 ```bash
+# terminal A
 bash services/start_backend.sh
+
+# terminal B
 bash services/start_frontend.sh
 bash services/start_skill_agent.sh
 ```
+
+If frontend starts before backend, the UI now shows a clear backend-not-reachable hint (instead of only `Load failed`).
 
 Behavior:
 - If backend Python deps are installed, backend runs in full mode.
@@ -85,7 +90,7 @@ Generated files:
 - `query_catalog.json`
 - `semantic_catalog.json` (includes `dsl_filters`)
 - `chart_plan.json`
-- `dashboard.json` (includes page-level `global_filters`)
+- `dashboard.json` (includes page-level `global_filters`, plus default `ui` unless `--without-ui-theme` is used)
 - `ui/` static scaffold
 - `services/` generated service bundle
 
@@ -111,9 +116,10 @@ Query data API now executes stored SQL against configured datasource (read-only 
 Example import:
 
 ```bash
+# Use an absolute path; replace <repo_root> with your local clone path.
 curl -X POST http://127.0.0.1:18000/api/v1/import/sql-md \
   -H 'Content-Type: application/json' \
-  -d '{"sql_md_path":"/Users/lyg/software/sql2bi/testdata/sql/demo.sql.md"}'
+  -d '{"sql_md_path":"<repo_root>/testdata/sql/demo.sql.md"}'
 ```
 
 ## Integration Tests
@@ -176,6 +182,19 @@ bash scripts/install-git-hooks.sh
 Hooks:
 - `pre-commit`: conflict markers, shell/python syntax, requirement pin checks
 - `pre-push`: smoke pipeline run on `sample.sql.md`
+
+## Unit Tests (minimal)
+
+Run script-level unit tests:
+
+```bash
+python3 -m unittest discover -s tests -p 'test_*.py'
+```
+
+Current coverage focus:
+- `skills/sql-to-bi-builder/scripts/build_dashboard_spec.py`
+  - helper behavior (`unique_keep_order`, `next_position`)
+  - CLI toggle behavior (`--with-ui-theme` / `--without-ui-theme`)
 
 ## Project Intro Website (Free)
